Fixed variable types

+++ backend/src/main/java/COMP_49X_our_search/backend/database/repositories/DepartmentRepository.java
@@ -3,4 +3,4 @@ package COMP_49X_our_search.backend.database.repositories;
 import COMP_49X_our_search.backend.database.entities.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentRepository extends JpaRepository<Department, Long> {}
+public interface DepartmentRepository extends JpaRepository<Department, Integer> {}
