+++ backend/src/main/java/COMP_49X_our_search/backend/database/repositories/ProjectRepository.java
@@ -4,7 +4,7 @@ import COMP_49X_our_search.backend.database.entities.Project;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepository extends JpaRepository<Project, Long> {
-  List<Project> findAllByDepartments_Id(Long departmentId);
-  List<Project> findAllByMajors_Id(Long majorId);
+public interface ProjectRepository extends JpaRepository<Project, Integer> {
+  List<Project> findAllByDepartments_Id(Integer departmentId);
+  List<Project> findAllByMajors_Id(Integer majorId);
 }
